requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cfgv==3.3.1
distlib==0.3.6
Django==4.1
django-bootstrap4==22.2
django-cleanup==6.0.0
django-colorfield==0.7.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.6.0
django-phonenumber-field==7.0.0
filelock==3.8.0
flake8==5.0.4
identify==2.5.3
isort==5.10.1
mccabe==0.7.0
nodeenv==1.7.0
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
PyYAML==6.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.16.4
